requirements.txt
PySide6>=6.0.0
python-osc>=1.8.0
python-dotenv>=1.0.0